CommonNitriteFilter created and pull down some methods from filter interface
