+++ nitrite-mvstore-adapter/src/test/java/org/dizitart/no2/integration/repository/RepositorySearchTest.java
@@ -40,9 +40,11 @@ import java.util.List;
 
 import static org.dizitart.no2.collection.FindOptions.orderBy;
 import static org.dizitart.no2.collection.FindOptions.skipBy;
-import static org.dizitart.no2.filters.Filter.*;
+import static org.dizitart.no2.filters.NitriteCommonFilters.*;
 import static org.dizitart.no2.filters.FluentFilter.$;
 import static org.dizitart.no2.filters.FluentFilter.where;
+import static org.dizitart.no2.filters.NitriteCommonFilters.and;
+import static org.dizitart.no2.filters.NitriteCommonFilters.or;
 import static org.junit.Assert.*;
 
 /**
