+++ nitrite-rocksdb-adapter/src/test/java/org/dizitart/no2/integration/repository/RepositorySearchTest.java
@@ -40,7 +40,7 @@ import java.util.List;
 
 import static org.dizitart.no2.collection.FindOptions.orderBy;
 import static org.dizitart.no2.collection.FindOptions.skipBy;
-import static org.dizitart.no2.filters.Filter.*;
+import static org.dizitart.no2.filters.NitriteCommonFilters.*;
 import static org.dizitart.no2.filters.FluentFilter.$;
 import static org.dizitart.no2.filters.FluentFilter.where;
 import static org.junit.Assert.*;
