+++ nitrite-spatial/src/test/java/org/dizitart/no2/spatial/SpatialIndexTest.java
@@ -38,7 +38,7 @@ import java.util.stream.Collectors;
 
 import static org.dizitart.no2.collection.Document.createDocument;
 import static org.dizitart.no2.collection.FindOptions.orderBy;
-import static org.dizitart.no2.filters.Filter.and;
+import static org.dizitart.no2.filters.NitriteCommonFilters.and;
 import static org.dizitart.no2.index.IndexType.UNIQUE;
 import static org.dizitart.no2.spatial.SpatialFluentFilter.where;
 import static org.dizitart.no2.spatial.SpatialIndexer.SPATIAL_INDEX;
