+++ nitrite/src/test/java/org/dizitart/no2/integration/collection/CollectionCompoundIndexTest.java
@@ -34,7 +34,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import static org.awaitility.Awaitility.await;
 import static org.dizitart.no2.collection.Document.createDocument;
-import static org.dizitart.no2.filters.Filter.and;
+import static org.dizitart.no2.filters.NitriteCommonFilters.and;
 import static org.dizitart.no2.filters.FluentFilter.where;
 import static org.dizitart.no2.index.IndexOptions.indexOptions;
 import static org.junit.Assert.*;
