+++ nitrite/src/test/java/org/dizitart/no2/integration/collection/CollectionFindTest.java
@@ -39,9 +39,10 @@ import static org.dizitart.no2.collection.Document.createDocument;
 import static org.dizitart.no2.collection.FindOptions.*;
 import static org.dizitart.no2.common.Constants.*;
 import static org.dizitart.no2.common.util.DocumentUtils.isSimilar;
-import static org.dizitart.no2.filters.Filter.*;
+import static org.dizitart.no2.filters.NitriteCommonFilters.*;
 import static org.dizitart.no2.filters.FluentFilter.$;
 import static org.dizitart.no2.filters.FluentFilter.where;
+import static org.dizitart.no2.filters.NitriteCommonFilters.*;
 import static org.dizitart.no2.integration.TestUtil.isSorted;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
